requirements.txt
streamlit
pandas
pandas-gbq
google-cloud-storage
google-auth-oauthlib
scikit-learn==1.3.2
xgboost==2.0.3
joblib
pyarrow
dill
geopy